Basic system for LIVE one-time project setup tasks
LIVE tasks are those that must be performed within Django or while Django is active.

Example: batch adding database entries
