+++ IFC/templates/IFC/login.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <!-- Required meta tags -->
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+
+    <title>Log in</title>
+    
+    {% load static %}
+    <link rel="icon" type="image/x-icon" href="{% static 'IFC/images/favicon.png' %}" />
+    
+</head>
+
+<body>
+    <!-- Navbar -->
+    {% include 'IFC/navbar.html' %}
+
+    
+    
+    <!-- Footer -->
+    {% include 'IFC/footer.html' %}
+</body>
+
+</html>
