--- IFC/templates/IFC/register.html
@@ -1,54 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <!-- Required meta tags -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-
-    <title>Calendar</title>
-    
-    {% load static %}
-    <link rel="icon" type="image/x-icon" href="{% static 'IFC/images/favicon.png' %}" />
-    
-
-    <style>
-        body {
-            margin: 0;
-            padding: 0;
-        }
-        
-        .container {
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            min-height: calc(100vh - 100px); /* Adjust the height as needed */
-        }
-        
-        .red-bar{
-            background-color: rgb(214, 0, 28);
-            color: #000;
-            text-align: center;
-            padding: 15px;
-        }
-        .page-title{
-            color: #000;
-            text-align: center;
-            padding-top: 25px;
-            padding-bottom: 0px;
-        }
-    </style>
-</head>
-
-<body>
-    <!-- Navbar -->
-    {% include 'IFC/navbar.html' %}
-
-    <h1 class="red-bar"></h1>
-    <h2 class = "page-title">Recruitment Registration</h2>
-    
-    <!-- Footer -->
-    {% include 'IFC/footer.html' %}
-</body>
-
-</html>
